Split off a micrometer-instrumentation module from core
The micrometer-core module currently contains both:
1) necessary classes for third-parties to write instrumentation using micrometer
and
2) micrometer-provided instrumentation

The new introduced micrometer-instrumentation module aims to be the module for the latter. This would allow other projects to depend on micrometer-core without instrumentation of unrelated things included in the dependency.
The module name does not clearly indicate the distinction, but I failed to come up with a better name so far.